Create FindMedianFromDatastream
https://leetcode.com/problems/find-median-from-data-stream/

This approach uses 2 heaps. One heap is used to store the first half of the array and this is a maxheap.
The other heap is a minheap and is used to store the second half of the array.
Then both the heaps are rebalanced such that their lengths do not differ by at most 1. This is done by popping the root elements from the heap with greater length and pushing in the heap with lesser length until the difference in their lengths is at most 1.
Then, to find the median of the array, there are 2 checks:
if the lengths of both the heaps are equal, then we return maxheap[0] + minheap[0] / 2 (even length array)
otherwise we return the root of maxheap (odd length array)
The overall TC is o(logn)